+++ src/command_modules/azure-cli-storage/HISTORY.rst
@@ -9,6 +9,7 @@ Release History
 * Make '--resource-group' parameter optional for 'storage account' commands.
 * Remove 'Failed precondition' warnings for individual failures in batch commands for single aggregated message.
 * blob/file delete-batch commands no longer output array of nulls.
+* blob download/upload/delete-batch commands will read sas-token from container url
 
 2.1.1
 +++++